lint: rename ClickableAddress to CopyableAddress, add keys to seller addresses list
