+++ src-gui/src/renderer/components/other/InlineCode.tsx
@@ -13,7 +13,8 @@ const useStyles = makeStyles(theme => ({
         alignItems: 'center',
         backgroundColor: theme.palette.grey[900],
         borderRadius: theme.shape.borderRadius,
-        padding: theme.spacing(0.5)
+        padding: theme.spacing(1),
+        gap: theme.spacing(1)
     }
 }))
